# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel Survey API
  description: |
    A comprehensive survey management API built with Laravel, featuring:
    - Multi-role authentication (Admin, User, LINE OA)
    - Real-time survey updates via WebSocket broadcasting
    - Advanced analytics and reporting
    - Export/Import functionality for surveys and responses
    - Tiered rate limiting for security
    - Email verification and password reset
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use JWT token obtained from authentication endpoints
  schemas:
...
    Survey:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Customer Satisfaction Survey"
        description:
          type: string
          nullable: true
          example: "Please rate your experience with our service"
        status:
          type: string
          enum: [draft, active, closed]
          example: "active"
        created_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
        questions:
          type: array
          items:
            $ref: '#/components/schemas/SurveyQuestion'
    SurveyQuestion:
      type: object
      properties:
        id:
          type: integer
          example: 1
        survey_id:
          type: integer
          example: 1
        question:
          type: string
          example: "How satisfied are you with our service?"
        type:
          type: string
          enum: [text, radio, checkbox, select, textarea, number]
          example: "radio"
        options:
          type: string
          nullable: true
          example: "Very Satisfied,Satisfied,Neutral,Unsatisfied,Very Unsatisfied"
        required:
          type: boolean
          example: true
        order:
          type: integer
          example: 1
        created_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-07-21T10:30:00Z"
...
    SurveyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "Customer Satisfaction Survey"
        description:
          type: string
          nullable: true
          example: "Please rate your experience"
        status:
          type: string
          enum: [draft, active, closed]
          example: "draft"
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation error"
        errors:
          type: object
          nullable: true
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "The given data was invalid."
        errors:
          type: object
          example:
            email: ["The email field is required."]
            password: ["The password field is required."]
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        links:
          type: object
          properties:
            first:
              type: string
              example: "http://localhost:8000/api/v1/admin/surveys?page=1"
            last:
              type: string
              example: "http://localhost:8000/api/v1/admin/surveys?page=5"
            prev:
              type: string
              nullable: true
              example: null
            next:
              type: string
              nullable: true
              example: "http://localhost:8000/api/v1/admin/surveys?page=2"
        meta:
          type: object
          properties:
            current_page:
              type: integer
              example: 1
            from:
              type: integer
              example: 1
            last_page:
              type: integer
              example: 5
            per_page:
              type: integer
              example: 15
            to:
              type: integer
              example: 15
            total:
              type: integer
              example: 75

    # Additional Request Models
...
    DashboardAnalytics:
      type: object
      properties:
        total_surveys:
          type: integer
          example: 25
        active_surveys:
          type: integer
          example: 8
        total_responses:
          type: integer
          example: 1250
        total_users:
          type: integer
          example: 150
        response_rate:
          type: number
          format: float
          example: 75.5
        recent_activity:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                example: "survey_created"
              message:
                type: string
                example: "New survey 'Customer Feedback' created"
              timestamp:
                type: string
                format: date-time
                example: "2024-07-21T10:30:00Z"
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Unauthenticated"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Forbidden"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Resource not found"
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Too Many Requests"
...
paths:
...
  /v1/admin/analytics/dashboard:
    get:
      tags:
        - Admin - Analytics
      summary: Get dashboard analytics
      description: Retrieve overall analytics data for admin dashboard
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Dashboard analytics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DashboardAnalytics'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...
  /v1/surveys:
    get:
      tags:
        - Surveys
      summary: List active surveys
      description: Get list of active surveys available to public
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - name: per_page
          in: query
          schema:
            type: integer
            default: 15
            maximum: 100
          description: Number of items per page
      responses:
        '200':
          description: Surveys retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Survey'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /v1/surveys/{survey}:
    get:
      tags:
        - Surveys
      summary: Get survey details
      description: Retrieve specific survey with questions
      parameters:
        - name: survey
          in: path
          required: true
          schema:
            type: integer
          description: Survey ID
      responses:
        '200':
          description: Survey details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Survey'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'

  # Survey Response Endpoints
...
  /v1/admin/surveys:
    get:
      tags:
        - Admin - Surveys
      summary: List all surveys (Admin)
      description: Get paginated list of all surveys for administrators
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number for pagination
        - name: per_page
          in: query
          schema:
            type: integer
            default: 15
            maximum: 100
          description: Number of items per page
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, closed]
          description: Filter by survey status
        - name: search
          in: query
          schema:
            type: string
          description: Search surveys by name or description
      responses:
        '200':
          description: Surveys retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Survey'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'

    post:
      tags:
        - Admin - Surveys
      summary: Create new survey
      description: Create a new survey
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SurveyRequest'
      responses:
        '201':
          description: Survey created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Survey created successfully"
                  data:
                    $ref: '#/components/schemas/Survey'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
...